third_party/WebKit/LayoutTests/fast/dom/window-polluter-tricky.html

<script src="../js/resources/js-test-pre.js"></script>
<form name="Taco" id="Taco"></form>
<applet name="Bravo" id="Bravo"></applet>
<script>
shouldBe("window.Taco", "document.getElementById('Taco')");
shouldBe("window.Bravo", "document.getElementById('Bravo')");

shouldBe("document.Taco", "document.getElementById('Taco')");
shouldBe("document.Bravo", "document.getElementById('Bravo')");
</script>
<script src="../js/resources/js-test-post.js"></script>
